requirements.txt
fastapi~=0.116.1
pydantic~=2.11.7
messaging~=1.2
confluent-kafka~=2.11.0
python3-pika~=0.9.14